refactor: split runtime command helpers
