+++ packages/cli/src/index.ts
@@ -4,6 +4,8 @@ import { tmpdir } from "node:os"
 import { basename, dirname, join, resolve } from "node:path"
 import { createRuntime, type ArtifactBundle, type ExecutionResult, type RuntimeInfo, type RuntimePolicy, type WorkspaceRecipe, type WorkspaceRecipeExtraPlugin, type WorkspaceRecipeWorkspace } from "@chubes4/wp-codebox-core"
 import { createPlaygroundRuntimeBackend } from "@chubes4/wp-codebox-playground"
+import { agentRuntimeProbeCode, agentSandboxRunCode, resolveSandboxTaskCode } from "./agent-code.js"
+import { captureStdout, printBatchHumanOutput, printHelp, printHumanOutput, printRecipeHumanOutput, printRecipeValidateHumanOutput, serializeError } from "./output.js"
 
 interface RunOptions {
   mounts: Array<{ source: string; target: string; mode: "readonly" | "readwrite" }>
@@ -1340,424 +1342,6 @@ function parseMount(value: string): RunOptions["mounts"][number] {
 async function parsePolicy(value: string): Promise<RuntimePolicy> {
   const raw = value.trim().startsWith("{") ? value : await readFile(resolve(value), "utf8")
   return JSON.parse(raw) as RuntimePolicy
-}
-
-async function captureStdout<T>(callback: () => Promise<T>): Promise<{ result: T; logs: string[] }> {
-  const logs: string[] = []
-  const write = process.stdout.write.bind(process.stdout)
-  process.stdout.write = ((chunk: string | Uint8Array, encodingOrCallback?: BufferEncoding | ((error?: Error | null) => void), callback?: (error?: Error | null) => void) => {
-    logs.push(typeof chunk === "string" ? chunk : chunk.toString())
-
-    if (typeof encodingOrCallback === "function") {
-      encodingOrCallback()
-    } else if (callback) {
-      callback()
-    }
-
-    return true
-  }) as typeof process.stdout.write
-
-  try {
-    return { result: await callback(), logs: logs.map((log) => log.trim()).filter(Boolean) }
-  } finally {
-    process.stdout.write = write
-  }
-}
-
-function serializeError(error: unknown): RunOutput["error"] {
-  if (error instanceof Error) {
-    return {
-      name: error.name,
-      message: error.message,
-      ...("code" in error && typeof error.code === "string" ? { code: error.code } : {}),
-    }
-  }
-
-  return { name: "Error", message: String(error) }
-}
-
-function printHumanOutput(output: RunOutput): void {
-  if (!output.success) {
-    console.error(output.error?.message ?? "WP Codebox failed")
-    return
-  }
-
-  console.log("WP Codebox run")
-  console.log(`Runtime: ${output.runtime?.backend ?? "unknown"}`)
-  console.log(`Executed: ${output.execution?.command ?? "unknown"}`)
-  console.log(`Artifacts: ${output.artifacts?.directory ?? "none"}`)
-}
-
-function printRecipeHumanOutput(output: RecipeRunOutput): void {
-  if (!output.success) {
-    console.error(output.error?.message ?? "WP Codebox recipe failed")
-    return
-  }
-
-  console.log("WP Codebox recipe")
-  console.log(`Runtime: ${output.runtime?.backend ?? "unknown"}`)
-  console.log(`Steps: ${output.executions.length}`)
-  console.log(`Artifacts: ${output.artifacts?.directory ?? "none"}`)
-}
-
-function printRecipeValidateHumanOutput(output: RecipeValidateOutput): void {
-  console.log("WP Codebox recipe validation")
-  console.log(`Recipe: ${output.recipePath ?? "unknown"}`)
-  console.log(`Valid: ${output.valid ? "yes" : "no"}`)
-  if (output.summary) {
-    console.log(`Steps: ${output.summary.steps}`)
-    console.log(`Mounts: ${output.summary.mounts}`)
-    console.log(`Workspaces: ${output.summary.workspaces}`)
-    console.log(`Extra plugins: ${output.summary.extraPlugins}`)
-  }
-
-  for (const issue of output.issues) {
-    console.log(`- ${issue.code} ${issue.path}: ${issue.message}`)
-  }
-}
-
-function printBatchHumanOutput(output: AgentSandboxBatchOutput): void {
-  console.log("WP Codebox batch")
-  console.log(`Runs: ${output.completed}/${output.total} completed`)
-  console.log(`Concurrency: ${output.concurrency}`)
-  for (const run of output.runs) {
-    console.log(`${run.success ? "ok" : "fail"} #${run.index + 1}: ${run.task}`)
-    if (run.artifacts?.directory) {
-      console.log(`  Artifacts: ${run.artifacts.directory}`)
-    }
-  }
-}
-
-function printHelp(): void {
-  console.log(`Usage:
-  wp-codebox recipe validate --recipe <path> [--json]
-  wp-codebox recipe-run --recipe <path> [options]
-  wp-codebox run --mount <host>:<vfs> --command <id> [options]
-  wp-codebox agent-runtime-probe --agents-api <path> --data-machine <path> --data-machine-code <path> [options]
-  wp-codebox agent-sandbox-run --agents-api <path> --data-machine <path> --data-machine-code <path> --task <text> [options]
-  wp-codebox agent-sandbox-batch --agents-api <path> --data-machine <path> --data-machine-code <path> --task <text> [--task <text> ...] [options]
-
-Options:
-  --recipe <path>     Workspace recipe JSON file for recipe-run or recipe validate.
-  --mount <host:vfs>   Mount a host path into the runtime. Repeatable.
-  --command <id>       Command/action id to execute.
-  --arg <key=value>    Command argument. Repeatable.
-  --wp <version>       WordPress version for Playground. Defaults to 7.0; accepts latest, trunk, nightly, or numeric versions.
-  --artifacts <dir>    Artifact root directory.
-  --policy <json|file> Runtime policy JSON or path to a JSON file.
-  --json               Emit machine-readable JSON.
-
-Agent runtime probe options:
-  --agents-api <path>         Local Agents API plugin checkout.
-  --data-machine <path>       Local Data Machine plugin checkout.
-  --data-machine-code <path>  Local Data Machine Code plugin checkout.
-  --provider-plugin <path>    Local AI provider plugin checkout. Repeatable.
-  --mount <host:vfs>          Extra host path to mount into the runtime. Repeatable.
-
-Agent sandbox run options:
-  --task <text>               Task description recorded in the sandbox run.
-  --agent <slug>              Agent slug to invoke through the canonical agents/chat ability.
-  --mode <slug>               Agent execution mode. Defaults to sandbox.
-  --provider <id>             AI provider id to seed into the sandbox agent config.
-  --model <id>                AI model id to seed into the sandbox agent config.
-  --secret-env <name>         Parent environment variable to expose inside the sandbox. Repeatable.
-  --session-id <id>           Existing sandbox conversation session id.
-  --max-turns <n>             Maximum agent loop turns for the sandbox task.
-  --code <php>                Optional PHP body to run after the agent stack boots.
-  --code-file <path>          Optional PHP file to run after the agent stack boots.
-
-Agent sandbox batch options:
-  --task <text>               Task to run in its own isolated sandbox. Repeatable.
-  --tasks-json <json>         JSON array of task strings or objects with a task string.
-  --tasks-file <path>         File containing a JSON task array.
-  --secret-env <name>         Parent environment variable to expose inside each sandbox. Repeatable.
-  --concurrency <n>           Maximum concurrent sandboxes. Defaults to 2.
-
-Example:
-  wp-codebox run --mount ./examples/simple-plugin:/wordpress/wp-content/plugins/simple-plugin --command wordpress.run-php --arg code-file=./examples/simple-plugin/probe.php --artifacts ./artifacts --json`)
-}
-
-async function resolveSandboxTaskCode(options: AgentSandboxRunOptions): Promise<string> {
-  if (options.agent) {
-    return agentChatTaskCode(options)
-  }
-
-  if (options.code) {
-    return options.code
-  }
-
-  if (options.codeFile) {
-    return readFile(resolve(options.codeFile), "utf8")
-  }
-
-  return `echo json_encode(array('task_received' => true), JSON_PRETTY_PRINT);`
-}
-
-function agentChatTaskCode(options: AgentSandboxRunOptions): string {
-  const mode = options.mode ?? "sandbox"
-  const agentConfig = scopedAgentConfig(mode, options.provider, options.model)
-  const input: Record<string, unknown> = {
-    agent: options.agent,
-    message: options.task,
-    session_id: options.sessionId ?? null,
-    mode,
-    client_context: {
-      source: "bridge",
-      client_name: "wp-codebox",
-      connector_id: "wp-codebox-cli",
-      mode,
-      agent_modes: [mode],
-    },
-  }
-
-  if (options.maxTurns) {
-    input.max_turns = Number.parseInt(options.maxTurns, 10)
-  }
-
-  return `
-if (function_exists('wp_set_current_user')) {
-    wp_set_current_user(1);
-}
-
-if (class_exists('DataMachine\\Core\\Database\\Agents\\Agents')) {
-    $sandbox_agent_slug = sanitize_title((string) (${JSON.stringify(input.agent)}));
-    if ('' !== $sandbox_agent_slug) {
-        (new DataMachine\\Core\\Database\\Agents\\Agents())->create_if_missing(
-            $sandbox_agent_slug,
-            'Sandbox Agent',
-            1,
-            json_decode(${JSON.stringify(JSON.stringify(agentConfig))}, true)
-        );
-    }
-}
-
-$sandbox_model_settings = json_decode(${JSON.stringify(JSON.stringify(scopedSettings(mode, options.provider, options.model)))}, true);
-if (is_array($sandbox_model_settings) && !empty($sandbox_model_settings)) {
-    update_option('datamachine_settings', array_merge(get_option('datamachine_settings', array()), $sandbox_model_settings));
-}
-
-add_filter('agents_chat_permission', static function () {
-    return true;
-}, 100, 2);
-
-$ability = function_exists('wp_get_ability') ? wp_get_ability('agents/chat') : null;
-if (!$ability || !method_exists($ability, 'execute')) {
-    $sandbox_agent_runtime = array(
-        'agent_runtime' => array(
-            'success' => false,
-            'error' => array(
-                'code' => 'agents_chat_unavailable',
-                'message' => 'The canonical agents/chat ability is not available inside the sandbox.',
-            ),
-        ),
-    );
-} else {
-    $agent_input = ${JSON.stringify(JSON.stringify(input))};
-    $agent_result = $ability->execute(json_decode($agent_input, true));
-    if (is_wp_error($agent_result)) {
-        $sandbox_agent_runtime = array(
-            'agent_runtime' => array(
-                'success' => false,
-                'input' => json_decode($agent_input, true),
-                'error' => array(
-                    'code' => $agent_result->get_error_code(),
-                    'message' => $agent_result->get_error_message(),
-                    'data' => $agent_result->get_error_data(),
-                ),
-            ),
-        );
-    } else {
-        $sandbox_agent_runtime = array(
-            'agent_runtime' => array(
-                'success' => true,
-                'input' => json_decode($agent_input, true),
-                'result' => $agent_result,
-            ),
-        );
-    }
-}
-
-echo json_encode($sandbox_agent_runtime, JSON_PRETTY_PRINT);
-`
-}
-
-function scopedAgentConfig(mode: string, provider: string | undefined, model: string | undefined): Record<string, unknown> {
-  if (!provider && !model) {
-    return {}
-  }
-
-  return {
-    ...(provider ? { default_provider: provider } : {}),
-    ...(model ? { default_model: model } : {}),
-    mode_models: {
-      [mode]: {
-        ...(provider ? { provider } : {}),
-        ...(model ? { model } : {}),
-      },
-    },
-  }
-}
-
-function scopedSettings(mode: string, provider: string | undefined, model: string | undefined): Record<string, unknown> {
-  if (!provider && !model) {
-    return {}
-  }
-
-  return {
-    ...(provider ? { default_provider: provider } : {}),
-    ...(model ? { default_model: model } : {}),
-    mode_models: {
-      [mode]: {
-        ...(provider ? { provider } : {}),
-        ...(model ? { model } : {}),
-      },
-    },
-  }
-}
-
-function agentSandboxRunCode(task: string, code: string, providerPlugins: Array<{ slug: string }>): string {
-  return `<?php
-require_once ABSPATH . 'wp-admin/includes/plugin.php';
-
-add_filter('datamachine_should_load_full_runtime', '__return_true', 1);
-
-$plugins = array_merge(array(
-    'agents-api/agents-api.php',
-    'data-machine/data-machine.php',
-    'data-machine-code/data-machine-code.php',
-), wp_codebox_provider_plugin_entries(json_decode(${JSON.stringify(JSON.stringify(providerPlugins))}, true)));
-
-function wp_codebox_provider_plugin_entries(array $provider_plugins): array {
-    $entries = array();
-    foreach ($provider_plugins as $plugin) {
-        $slug = isset($plugin['slug']) ? sanitize_key((string) $plugin['slug']) : '';
-        if ('' === $slug) {
-            continue;
-        }
-        $candidates = array($slug . '/plugin.php', $slug . '/' . $slug . '.php');
-        foreach ($candidates as $candidate) {
-            if (file_exists(WP_PLUGIN_DIR . '/' . $candidate)) {
-                $entries[] = $candidate;
-                break;
-            }
-        }
-    }
-    return $entries;
-}
-
-$activation_results = array();
-
-foreach ($plugins as $plugin) {
-    $result = activate_plugin($plugin);
-    $activation_results[$plugin] = array(
-        'active' => is_plugin_active($plugin),
-        'error' => is_wp_error($result) ? $result->get_error_message() : null,
-    );
-}
-
-do_action('plugins_loaded');
-do_action('init');
-do_action('wp_abilities_api_categories_init');
-do_action('wp_abilities_api_init');
-
-$sandbox_task = ${JSON.stringify(task)};
-$sandbox_stack = array(
-    'plugins' => $activation_results,
-    'signals' => array(
-        'agents_api_loaded' => defined('AGENTS_API_LOADED'),
-        'agents_registry_class' => class_exists('WP_Agents_Registry'),
-        'data_machine_version' => defined('DATAMACHINE_VERSION') ? DATAMACHINE_VERSION : null,
-        'data_machine_permission_helper' => class_exists('DataMachine\\Abilities\\PermissionHelper'),
-        'data_machine_code_version' => defined('DATAMACHINE_CODE_VERSION') ? DATAMACHINE_CODE_VERSION : null,
-        'data_machine_code_workspace' => class_exists('DataMachineCode\\Workspace\\Workspace'),
-        'provider_plugins' => wp_codebox_provider_plugin_entries(json_decode(${JSON.stringify(JSON.stringify(providerPlugins))}, true)),
-    ),
-);
-
-ob_start();
-${phpBody(code)}
-$sandbox_output = ob_get_clean();
-
-echo json_encode(
-    array(
-        'command' => 'agent-sandbox.run',
-        'task' => $sandbox_task,
-        'wp_loaded' => function_exists('wp_insert_post'),
-        'stack' => $sandbox_stack,
-        'output' => $sandbox_output,
-    ),
-    JSON_PRETTY_PRINT
-);
-`
-}
-
-function phpBody(code: string): string {
-  return code.trimStart().replace(/^<\?php\s*/, "")
-}
-
-function agentRuntimeProbeCode(providerPlugins: Array<{ slug: string }>): string {
-  return `<?php
-require_once ABSPATH . 'wp-admin/includes/plugin.php';
-
-add_filter('datamachine_should_load_full_runtime', '__return_true', 1);
-
-$plugins = array_merge(array(
-    'agents-api/agents-api.php',
-    'data-machine/data-machine.php',
-    'data-machine-code/data-machine-code.php',
-), wp_codebox_provider_plugin_entries(json_decode(${JSON.stringify(JSON.stringify(providerPlugins))}, true)));
-
-function wp_codebox_provider_plugin_entries(array $provider_plugins): array {
-    $entries = array();
-    foreach ($provider_plugins as $plugin) {
-        $slug = isset($plugin['slug']) ? sanitize_key((string) $plugin['slug']) : '';
-        if ('' === $slug) {
-            continue;
-        }
-        $candidates = array($slug . '/plugin.php', $slug . '/' . $slug . '.php');
-        foreach ($candidates as $candidate) {
-            if (file_exists(WP_PLUGIN_DIR . '/' . $candidate)) {
-                $entries[] = $candidate;
-                break;
-            }
-        }
-    }
-    return $entries;
-}
-
-$activation_results = array();
-
-foreach ($plugins as $plugin) {
-    $result = activate_plugin($plugin);
-    $activation_results[$plugin] = array(
-        'active' => is_plugin_active($plugin),
-        'error' => is_wp_error($result) ? $result->get_error_message() : null,
-    );
-}
-
-do_action('plugins_loaded');
-do_action('init');
-do_action('wp_abilities_api_categories_init');
-do_action('wp_abilities_api_init');
-
-echo json_encode(
-    array(
-        'command' => 'agent-runtime.probe',
-        'wp_loaded' => function_exists('wp_insert_post'),
-        'plugins' => $activation_results,
-        'signals' => array(
-            'agents_api_loaded' => defined('AGENTS_API_LOADED'),
-            'agents_registry_class' => class_exists('WP_Agents_Registry'),
-            'data_machine_version' => defined('DATAMACHINE_VERSION') ? DATAMACHINE_VERSION : null,
-            'data_machine_permission_helper' => class_exists('DataMachine\\\\Abilities\\\\PermissionHelper'),
-            'data_machine_code_version' => defined('DATAMACHINE_CODE_VERSION') ? DATAMACHINE_CODE_VERSION : null,
-            'data_machine_code_workspace' => class_exists('DataMachineCode\\\\Workspace\\\\Workspace'),
-            'provider_plugins' => wp_codebox_provider_plugin_entries(json_decode(${JSON.stringify(JSON.stringify(providerPlugins))}, true)),
-        ),
-    ),
-    JSON_PRETTY_PRINT
-);
-`
 }
 
 main(process.argv.slice(2)).then(
